requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
darkskylib==0.3.91
Django==2.1.11
fake-useragent==0.1.11
ipstack==0.1.3
useragent==0.1.1
lxml==4.4.1
python-dateutil==2.8.0
pandas==0.25.2
